Change directory in the debug buffer instead of expanding filenames
In the most likely case of the game files being relative paths, it is
now the relative paths which are passed as command line arguments to
the debugger.  This shortens error output in the debugger when error
messages print the filename.

When using @save in the debugger, the game input is now saved into the
game directory.