Clean-up parser and indentation code
Reclassify # as symbol syntax in the syntax table.

Always parse and store the statement contents in relevant parser
structs.

Don't track the prefix character for a statement because the value was
never used.

Treat (or) as a block open and close.

Move the special case of "tight bracing" (a block opening where a
non-whitespace character immediately follows the opening brace) to the
indentation code from the block parsing code.  Avoid dedenting a line
inside such a block.